% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create a package}
\usage{
create_package(
  path,
  use_git = TRUE,
  use_github = use_git,
  use_ci = use_github,
  use_precommit = use_ci,
  use_coverage = use_ci,
  use_lintr = TRUE,
  use_tests = TRUE,
  ...,
  open = FALSE
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{...}{Additional arguments passed to \code{usethis::create_package()}}

\item{open}{If \code{TRUE}, \link[usethis:proj_activate]{activates} the new project:
\itemize{
\item If RStudio desktop, the package is opened in a new session.
\item If on RStudio server, the current RStudio project is activated.
\item Otherwise, the working directory and active project is changed.
}}

\item{use.git}{Initialise and make initial commits to a local git
repository. By default is \code{TRUE}.}

\item{use.github}{Create and push to a GitHub repository. By default takes
value of \code{use.git}. If set to \code{TRUE}, \code{use.git} must also be \code{TRUE}.}

\item{use.ci}{Set up a GitHub Action job to run R CMD CHECK. By default takes
value of \code{use.github}. If set to \code{TRUE}, \code{use.github} must also be \code{TRUE}.}

\item{use.precommit}{Set up \code{precommit} to run checks before any commit. By
default takes value of \code{use.ci}. If set to \code{TRUE}, \code{use.git} must also
be \code{TRUE}.}

\item{use.coverage}{Set up \code{covr} and, if using GitHub, a GitHub Action to run
code coverage and use the Codecov service. By default takes value of \code{use.ci}.}

\item{use.lintr}{Set up lintr. By default is \code{TRUE}.}

\item{use.tests}{Set up testthat. By default is \code{TRUE}.}
}
\value{
The path to the created package, invisibly.
}
\description{
Creates a new package at given path; the package is named as the
last folder in the path. Currently there are no configuration options beyond
what is exposed by \code{usethis::createpackage()}.
}
