% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny.R
\name{create_shiny}
\alias{create_shiny}
\title{Create a \code{{shiny}} dashboard, with optional software development best practices}
\usage{
create_shiny(
  path,
  only_git = FALSE,
  use_git = TRUE,
  use_github = !only_git,
  use_ci = use_github,
  use_precommit = use_ci,
  use_coverage = use_ci,
  use_lintr = !only_git,
  use_tests = !only_git,
  open = FALSE,
  project_hook = NULL,
  ...
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{only_git}{Boolean. When \code{TRUE}, sets all flags except \code{use_git} to \code{FALSE}.
Default \code{FALSE}.}

\item{use_git}{Boolean. Initialise and make initial commits to a local git
repository. Default \code{TRUE}.}

\item{use_github}{Boolean. Create and push to a GitHub repository. By default
takes value of \code{use_git}. If set to \code{TRUE}, \code{use_git} must also be \code{TRUE}.}

\item{use_ci}{Boolean. Set up a GitHub Action job to run R CMD CHECK. By
default takes value of \code{!only_git}. If set to \code{TRUE}, \code{use_github} must
also be \code{TRUE}.}

\item{use_precommit}{Boolean. Set up \code{precommit} to run checks before any
commit. By default takes value of \code{use_ci}. If set to \code{TRUE}, \code{use_git} must
also be \code{TRUE}.}

\item{use_coverage}{Boolean. Set up \code{{covr}} and, if using GitHub, a GitHub
Action to run code coverage and use the Codecov service. By default takes
value of \code{use_ci}.}

\item{use_lintr}{Boolean. Set up lintr. Default \code{!only_git}.}

\item{use_tests}{Boolean. Set up testthat. Default \code{!only_git}.}

\item{open}{Boolean. Open the created project? Default \code{FALSE}.}

\item{project_hook}{A function executed as a hook after project
creation. Can be used to change the default \code{{golem}} structure.
to override the files and content. This function is executed just
after the project is created.}

\item{...}{Arguments passed to project_hook.}
}
\value{
Path to the newly created \code{{shiny}} package, invisibly.
}
\description{
Creates a new \code{{shiny}} dashboard as a  package at given path; the
package is named as the last folder in the path.
}
