% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_dashboard.R
\name{create_shiny_dashboard}
\alias{create_shiny_dashboard}
\title{Create a shiny dashboard}
\usage{
create_shiny_dashboard(
  path,
  use_git = TRUE,
  use_github = use_git,
  use_ci = use_github,
  use_precommit = use_ci,
  use_coverage = use_ci,
  use_lintr = TRUE,
  use_tests = TRUE,
  ...,
  open = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{use_git}{Initialise and make initial commits to a local git
repository. By default is \code{TRUE}.}

\item{use_github}{Create and push to a GitHub repository. By default takes
value of \code{use_git}. If set to \code{TRUE}, \code{use_git} must also be \code{TRUE}.}

\item{use_ci}{Set up a GitHub Action job to run R CMD CHECK. By default takes
value of \code{use_github}. If set to \code{TRUE}, \code{use_github} must also be \code{TRUE}.}

\item{use_precommit}{Set up \code{precommit} to run checks before any commit. By
default takes value of \code{use_ci}. If set to \code{TRUE}, \code{use_git} must also
be \code{TRUE}.}

\item{use_coverage}{Set up \code{covr} and, if using GitHub, a GitHub Action to run
code coverage and use the Codecov service. By default takes value of \code{use_ci}.}

\item{use_lintr}{Set up lintr. By default is \code{TRUE}.}

\item{use_tests}{Set up testthat. By default is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{golem::create_golem()}}

\item{open}{Boolean. Open the created project?}

\item{overwrite}{Boolean. Should the already existing project be overwritten ?}
}
\value{
The path to the created package, invisibly.
}
\description{
Creates a new shiny dashboard as a  package at given path; the
package is named as the last folder in the path.
}
