% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create a package}
\usage{
create_package(path, ..., open = FALSE)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{...}{Additional arguments passed to \code{usethis::create_package()}}

\item{open}{If \code{TRUE}, \link[usethis:proj_activate]{activates} the new project:
\itemize{
\item If RStudio desktop, the package is opened in a new session.
\item If on RStudio server, the current RStudio project is activated.
\item Otherwise, the working directory and active project is changed.
}}
}
\value{
The path to the created package, invisibly.
}
\description{
Creates a new package at given path; the package is named as the
last folder in the path. Currently there are no configuration options beyond
what is exposed by \code{usethis::createpackage()}.
}
